% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/issues.R
\name{new_issue}
\alias{assign_issue}
\alias{close_issue}
\alias{edit_issue}
\alias{new_issue}
\alias{reopen_issue}
\alias{unassign_issue}
\title{Post a new issue or edit one}
\usage{
new_issue(title, project, ...)

edit_issue(issue_id, project, ...)

close_issue(issue_id, project, ...)

reopen_issue(issue_id, project, ...)

assign_issue(issue_id, assignee_id = NULL, project, ...)

unassign_issue(issue_id, project, ...)
}
\arguments{
\item{title}{title of the issue}

\item{project}{project where the issue should be posted}

\item{...}{further parameters passed to the API call, may
contain description, asignee_id, milestone_id, labels, state_event (for edit_issue).}

\item{issue_id}{id of issue to edit  (projectwide iid, not gitlab API id)}

\item{assignee_id}{numeric id of users as returned in '/users/' API request}
}
\description{
Post a new issue or edit one

Post a new issue or edit one
}

