% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{get_commits}
\alias{get_commits}
\alias{get_diff}
\title{Get commits and diff from a project repository}
\usage{
get_commits(project, commit_sha = c(), ...)

get_diff(project, commit_sha, ...)
}
\arguments{
\item{project}{project name or id}

\item{commit_sha}{if not null, get only the commit with the specific hash; for
\code{get_diff} this must be specified}

\item{...}{passed on to \code{\link{gitlab}} API call, may contain
\code{ref_name} for specifying a branch or tag to list commits of}
}
\description{
Get commits and diff from a project repository
}

