% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{get_file}
\alias{get_file}
\title{Get a file from a gitlab repository}
\usage{
get_file(project, file_path, ref = "master", to_char = TRUE, ...)
}
\arguments{
\item{project}{name or id of project}

\item{file_path}{path to file}

\item{ref}{name of ref (commit branch or tag)}

\item{to_char}{flag if output should be converted to char; otherwise it is of class raw}

\item{...}{passed on to \code{\link{gitlab}}}
}
\description{
Get a file from a gitlab repository
}

