% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_headers.R
\name{gitlabr-deprecated}
\alias{gitlabr-deprecated}
\alias{gl_builds}
\alias{gl_ci_job}
\title{Deprecated functions}
\usage{
gl_builds(project, api_version = 4, ...)

gl_ci_job()
}
\arguments{
\item{project}{id (preferred way) or name of the project.
Not repository name.}

\item{api_version}{Since \code{gl_builds} is no longer working for GitLab API v4,
this must be set to "3" in order to avoid deprecation
warning and HTTP error.  It currently
default to "4" with deprecation message.´}

\item{...}{Parameters to the new function}
}
\value{
Warning for deprecated functions and
output depending on the superseeding function.
}
\description{
List of deprecated functions that will be removed in future versions.
}
\section{Details}{

\tabular{rl}{
\code{gl_builds} \tab in favour of \code{gl_pipelines} \cr
\code{gl_ci_job} \tab in favour of \code{use_gitlab_ci} \cr
}
}

