% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlabr-package.R
\docType{package}
\name{gitlabr-package}
\alias{gitlabr}
\alias{gitlabr-package}
\title{gitlabr: Access to the 'Gitlab' API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides R functions to access the API of the project and
    repository management web application 'GitLab'. For many common tasks
    (repository file access, issue assignment and status, commenting)
    convenience wrappers are provided, and in addition the full API can be
    used by specifying request locations. 'GitLab' is open-source software
    and can be self-hosted or used on <https://about.gitlab.com>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://statnmap.github.io/gitlabr/}
  \item Report bugs at \url{https://github.com/statnmap/gitlabr/issues}
}

}
\author{
\strong{Maintainer}: Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Authors:
\itemize{
  \item Jirka Lewandowski \email{jirka.lewandowski@wzb.eu}
}

}
\keyword{internal}
