% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlabr-package.R
\docType{package}
\name{gitlabr}
\alias{gitlabr}
\alias{gitlabr-package}
\title{Interface to gitlab API on high and low levels}
\description{
gitlabr R package
}
\details{
Interface to gitlab API on high and low levels
\tabular{ll}{
Package: \tab gitlabr\cr
Type: \tab Package\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\references{
\url{http://blog.points-of-interest.cc/}
}
\author{
Jirka Lewandowski \email{jirka.lewandowski@wzb.eu}
}
