% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_env.R
\name{gitlabr_options_set}
\alias{gitlabr_options_set}
\title{Set gitlabr options}
\usage{
gitlabr_options_set(key, value)
}
\arguments{
\item{key}{option name}

\item{value}{option value}
}
\value{
Used for side effect. Populates user \code{\link[=options]{options()}}
}
\description{
Set gitlabr options
}
\details{
Options accounted for by gitlabr:
\itemize{
\item \code{gitlabr.main}: Name of the main branch of your repository. Default to "main" in functions.
}
}
\examples{
# Principal branch is called "master"
gitlabr_options_set("gitlabr.main", "master")
# Go back to default option (default branch will be "main")
gitlabr_options_set("gitlabr.main", NULL)
}
