% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{gl_create_merge_request}
\alias{gl_create_merge_request}
\title{Create a merge request}
\usage{
gl_create_merge_request(project, source_branch, target_branch = "master",
  title, description, verb = httr::POST, ...)
}
\arguments{
\item{project}{name or id of project (not repository!)}

\item{source_branch}{name of branch to be merged}

\item{target_branch}{name of branch into which to merge}

\item{title}{title of the merge request}

\item{description}{description text for the merge request}

\item{verb}{is ignored, will always be forced to match the action the function name indicates}

\item{...}{passed on to \code{\link{gitlab}}. Might contain more fields documented in gitlab API doc.}
}
\description{
Create a merge request
}
