% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{gl_file_exists}
\alias{gl_file_exists}
\alias{gl_get_file}
\title{For \code{gl_file_exists} dots are passed on to \code{\link{gl_list_files}} and gitlab API call}
\usage{
gl_file_exists(project, file_path, ...)

gl_get_file(project, file_path, ref = "master", to_char = TRUE, ...)
}
\arguments{
\item{project}{name or id of project}

\item{file_path}{path to file}

\item{...}{passed on to \code{\link{gitlab}}}

\item{ref}{name of ref (commit branch or tag)}

\item{to_char}{flag if output should be converted to char; otherwise it is of class raw}
}
\description{
For \code{gl_file_exists} dots are passed on to \code{\link{gl_list_files}} and gitlab API call

Get a file from a gitlab repository
}
