% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{gl_get_group_id}
\alias{gl_get_group_id}
\title{Get a group id by name}
\usage{
gl_get_group_id(group_name, ...)
}
\arguments{
\item{group_name}{group name}

\item{...}{passed on to \code{\link[=gitlab]{gitlab()}}}
}
\value{
Integer. ID of the group if found.
}
\description{
Get a group id by name
}
\details{
Number of pages searched is limited to
(per_page =) 20 * (max_page =) 10 by default.
If the \code{group_name} is an old group lost
in a big repository (position > 200),
\code{gl_get_group_id()} may not find the group id.
}
\examples{
\dontrun{
gl_get_group_id("<<your-group-name>>")
}
}
