% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{gl_group_req}
\alias{gl_group_req}
\title{Create a group specific request}
\usage{
gl_group_req(group, ...)
}
\arguments{
\item{group}{The ID, name or URL-encoded path of the group}

\item{...}{passed on to \code{\link[=gl_get_group_id]{gl_get_group_id()}}}
}
\value{
A vector of character to be used
as request for functions involving groups
}
\description{
Prefixes the request location with "groups/:id/subgroups" and automatically
translates group names into ids
}
\examples{
\dontrun{
gl_group_req("test_group" = "<<your-group-id>>")
}
}
