% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{gl_list_groups}
\alias{gl_list_groups}
\alias{gl_list_sub_groups}
\title{List groups information}
\usage{
gl_list_groups(...)

gl_list_sub_groups(group_id, ...)
}
\arguments{
\item{...}{passed on to \code{\link[=gitlab]{gitlab()}}}

\item{group_id}{id of the group to list group from}
}
\value{
tibble of each group with corresponding information
}
\description{
List groups information
}
\examples{
\dontrun{
set_gitlab_connection(
  gitlab_url = "https://gitlab.com", 
  private_token = Sys.getenv("GITLAB_COM_TOKEN")
)
# List all groups
gl_get_groups(max_page = 1)
# List sub-groups of a group
gl_list_sub_groups(group_id = "<<group-id>>", max_page = 1)
}
}
