% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{gl_list_user_projects}
\alias{gl_list_user_projects}
\title{List user projects}
\usage{
gl_list_user_projects(user_id, ...)
}
\arguments{
\item{...}{passed on to \code{\link[=gitlab]{gitlab()}}}
}
\description{
List user projects
}
\examples{
\dontrun{
set_gitlab_connection(
  gitlab_url = "https://gitlab.com", 
  private_token = Sys.getenv("GITLAB_COM_TOKEN")
)
gl_list_user_projects(user_id = "<<user-id>>", max_page = 1)
}
}
