% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{gl_new_project}
\alias{gl_new_project}
\title{Create new project}
\usage{
gl_new_project(name, path, ...)
}
\arguments{
\item{name}{of the new project. The name of the new project. Equals path if not provided}

\item{path}{to the new project if name is not provided. Repository name for new project. Generated based on name if not provided (generated as lowercase with dashes).}

\item{...}{passed on to \code{\link{gitlab}} API call for "Create project"}
}
\description{
Create new project
}
\examples{
\dontrun{
my_gitlab <- gl_connection(
  gitlab_url = "https://gitlab.com",
  private_token = Sys.getenv("GITLAB_TOKEN"))
gl_new_project(name = "toto", gitlab_con = my_gitlab)
}
}
