% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{gl_repository}
\alias{gl_repository}
\alias{gl_list_files}
\alias{gl_file_exists}
\alias{gl_get_file}
\title{Access to repository files in GitLab}
\usage{
gl_repository(req = c("tree"), project, ...)

gl_list_files(project, ...)

gl_file_exists(project, file_path, ref, ...)

gl_get_file(
  project,
  file_path,
  ref = "master",
  to_char = TRUE,
  api_version = 4,
  ...
)
}
\arguments{
\item{req}{request to perform on repository (everything after '/repository/'
in GitLab API, as vector or part of URL)}

\item{project}{name or id of project (not repository!)}

\item{...}{passed on to \code{\link{gitlab}} API call}

\item{file_path}{path to file}

\item{ref}{name of ref (commit branch or tag)}

\item{to_char}{flag if output should be converted to char; otherwise it is of class raw}

\item{api_version}{a switch to force deprecated GitLab API v3 behavior. See details section "API version" of \code{\link{gl_connection}}}
}
\description{
Access to repository files in GitLab

For \code{gl_file_exists} dots are passed on to \code{\link{gl_list_files}} and GitLab API call

Get a file from a GitLab repository
}
\examples{
\dontrun{
# Set GitLab connection for examples
my_gitlab <- gl_connection(
 gitlab_url = "https://gitlab.com",
 private_token = Sys.getenv("GITLAB_COM_TOKEN"))
# Set the connection for the session
set_gitlab_connection(my_gitlab)

# Access repository
# _All files
gl_repository(project = <<your-project-id>>)
# _All contributors
gl_repository(project = <<your-project-id>>, "contributors")
# _List files
gl_list_files(project = <<your-project-id>>)
# _Get content of one file
gl_get_file(project = <<your-project-id>>, file_path = "README.md")
# _Test if file exists
gl_file_exists(project = <<your-project-id>>, file_path = "README.md", ref = "master")
}
}
