% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr_extensions.R
\name{iff}
\alias{iff}
\alias{iffn}
\title{Apply a function if and only if test is TRUE}
\usage{
iff(obj, test, fun, ...)

iffn(obj, test, fun, ...)
}
\arguments{
\item{obj}{object to apply test and fun to}

\item{test}{logical or function to apply to test}

\item{fun}{function to apply}

\item{...}{passed on to test}
}
\description{
otherwise return input value unchanged
}
\details{
iffn is ... if and only if test is FALSE
}

