% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilist_to_tibble.R
\name{multilist_to_tibble}
\alias{multilist_to_tibble}
\title{Modify a multilist from API JSON output to a level 1 tibble}
\usage{
multilist_to_tibble(the_list)
}
\arguments{
\item{the_list}{list of element as issued from a API REST call}
}
\value{
a tibble with columns as the names of the list
}
\description{
Modify a multilist from API JSON output to a level 1 tibble
}
\examples{
reprex <- list(
  list(a = 1, b = list("email1", "email2", "email3"), c = list("3")),
  list(a = 5, b = list("email1"), c = list("4")),
  list(a = 3, b = NULL, c = list("3", "2"))
)

multilist_to_tibble(reprex)
}
