% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr_extensions.R
\name{pipe_into}
\alias{pipe_into}
\title{Pipe into specific formal argument}
\usage{
pipe_into(x, param_name, fun, ...)
}
\arguments{
\item{x}{value to be piped into fun}

\item{param_name}{name of the argument that x should be assigned to}

\item{fun}{function}

\item{...}{further arguments for fun}
}
\description{
This rotates the order of the arguments such that the one named
in param_name comes first and then calls the function.
}

