% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{proj_req}
\alias{proj_req}
\title{Create a project specific request}
\usage{
proj_req(project, req, ...)
}
\arguments{
\item{project}{project name or id}

\item{req}{character vector of request location}

\item{...}{passed on to \code{\link{get_project_id}}}
}
\description{
Prefixes the request location with "project/:id" and automatically
translates project names into ids
}

