% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{repository}
\alias{list_files}
\alias{repository}
\title{Access to repository functions in Gitlab API}
\usage{
repository(req = c("tree"), project, ...)

list_files(...)
}
\arguments{
\item{req}{request to perform on repository (everything after '/repository/'
in gitlab API, as vector or part of URL)}

\item{project}{name or id of project (not repository!)}

\item{...}{passed on to \code{\link{gitlab}} API call, may include \code{path} argument for path}
}
\description{
Access to repository functions in Gitlab API
}

