% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_code.R
\name{update_gitlabr_code}
\alias{update_gitlabr_code}
\title{Convert gitlabr legacy code to 0.7 compatible}
\usage{
update_gitlabr_code(file, text = readLines(file), internal = FALSE)
}
\arguments{
\item{file}{file to read from/write to. Maybe NULL for input and return only}

\item{text}{lines of code to convert}

\item{internal}{whether to also replace names of internal functions}
}
\description{
CAUTION: This functions output/results should be checked manually before
committing the code, since it uses regular expression heuristically
to parse code and cannot guarantee complete and correct code replacement
}
