% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{list_branches}
\alias{create_branch}
\alias{delete_branch}
\alias{list_branches}
\title{List, create and delete branches}
\usage{
list_branches(project, verb = httr::GET, ...)

create_branch(project, branch_name, ref = "master", verb = httr::POST, ...)

delete_branch(project, branch_name, verb = httr::POST, ...)
}
\arguments{
\item{project}{name or id of project (not repository!)}

\item{verb}{is ignored, will always be forced to match the action the function name indicates}

\item{...}{passed on to \code{\link{gitlab}}}

\item{branch_name}{name of branch to create/delete}

\item{ref}{ref name of origin for newly created branch}
}
\description{
List, create and delete branches

List, create and delete branches

List, create and delete branches
}

