% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{gl_ci_job}
\alias{gl_ci_job}
\title{Define GitLab CI jobs content}
\usage{
gl_ci_job()
}
\value{
Creates the content of a .gitlab-ci.yml file as character.
}
\description{
Exploration of job content is deprecated as of 'gitlabr' 1.1.7.
Content of .gitlab-ci.yml file is now created using templates with
use_gitlab_ci(type = "check-coverage-pkgdown"). See \code{\link[=use_gitlab_ci]{use_gitlab_ci()}}.
}
\examples{
\dontrun{
# Deprecated
gl_ci_job()
}
}
\seealso{
\code{\link[=use_gitlab_ci]{use_gitlab_ci()}}
}
