% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{gl_get_project_id}
\alias{gl_get_project_id}
\title{Get a project id by name}
\usage{
gl_get_project_id(project_name, verb = httr::GET, auto_format = TRUE,
  ...)
}
\arguments{
\item{project_name}{project name}

\item{verb}{ignored; all calls with this function will have \code{\link{gitlab}}'s
default verb \code{httr::GET}}

\item{auto_format}{ignored}

\item{...}{passed on to \code{\link{gitlab}}}
}
\description{
Get a project id by name
}
