% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gl_list_issues}
\alias{gl_list_issues}
\alias{gl_get_issue}
\title{Get issues of a project or user}
\usage{
gl_list_issues(project = NULL, issue_id = NULL, verb = httr::GET, ...)

gl_get_issue(issue_id, project, ...)
}
\arguments{
\item{project}{project name or id, may be null for all issues created by user}

\item{issue_id}{optional issue id (projectwide iid, not gitlab API id)}

\item{verb}{ignored; all calls with this function will have \code{\link{gitlab}}'s
default verb \code{httr::GET}}

\item{...}{further parameters passed on to \code{\link{gitlab}}, may be
state, labels, issue id, ...}
}
\description{
Get issues of a project or user
}
\details{
\code{gl_get_issue} provides a wrapper with swapped arguments for convenience, esp. when
using a project connection
}
