% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{gl_list_projects}
\alias{gl_list_projects}
\title{List projects in Gitlab}
\usage{
gl_list_projects(...)
}
\arguments{
\item{...}{passed on to \code{\link{gitlab}}}
}
\description{
List projects in Gitlab
}
\examples{
\dontrun{
my_gitlab <- gl_connection(...) ## fill in login parameters
my_gitlab(gl_list_projects)
}
}
