% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{gl_proj_req}
\alias{gl_proj_req}
\title{Create a project specific request}
\usage{
gl_proj_req(project, req, ...)
}
\arguments{
\item{project}{project name or id}

\item{req}{character vector of request location}

\item{...}{passed on to \code{\link[=gl_get_project_id]{gl_get_project_id()}}}
}
\value{
A vector of character to be used as request for functions involving projects
}
\description{
Prefixes the request location with "project/:id" and automatically
translates project names into ids
}
\examples{
\dontrun{
gl_proj_req("test_project"<<your-project-id>>, req = "merge_requests")
}
}
