% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_and_repos.R
\name{gl_repository}
\alias{gl_repository}
\alias{gl_list_files}
\title{Access to repository functions in Gitlab API}
\usage{
gl_repository(req = c("tree"), project, ...)

gl_list_files(...)
}
\arguments{
\item{req}{request to perform on repository (everything after '/repository/'
in gitlab API, as vector or part of URL)}

\item{project}{name or id of project (not repository!)}

\item{...}{passed on to \code{\link{gitlab}} API call, may include \code{path} argument for path}
}
\description{
Access to repository functions in Gitlab API
}
