% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gl_to_issue_id}
\alias{gl_to_issue_id}
\title{Translate projectwide issue id to global gitlab API issue id}
\usage{
gl_to_issue_id(issue_id, project, ...)
}
\arguments{
\item{issue_id}{projectwide issue id (as seen by e.g. gitlab website users)}

\item{project}{project name or id}

\item{...}{passed on to \code{\link{gitlab}}}
}
\description{
Translate projectwide issue id to global gitlab API issue id
}
